+++ internal/project/duplicate.go
@@ -338,7 +338,7 @@ func (m *Manager) reassignHostPorts(ctx context.Context, proj *store.Project) er
 				}
 				return swap(&c.ConsolePort)
 			})
-		case store.ServiceRedis, store.ServiceMemcached, store.ServiceMailpit, store.ServiceRabbitMQ, store.ServiceMeilisearch, store.ServiceTypesense:
+		case store.ServiceRedis, store.ServiceMemcached, store.ServiceMailpit, store.ServiceRabbitMQ, store.ServiceMeilisearch, store.ServiceTypesense, store.ServiceOpenSearch:
 			err = editConfig(svc, func(c *runtime.ServiceConfig) error {
 				if err := swap(&c.HostPort); err != nil {
 					return err